feat: 优化outputDir为空时的逻辑

+++ packages/cli/src/utils/clean.ts
@@ -10,11 +10,14 @@ const __cwd = process.cwd()
  */
 export const cleanPost = (postPath: string) => {
   try {
+    if (!postPath) {
+      out.info('outputDir字段未配置')
+      return
+    }
     const dist = path.join(__cwd, postPath)
     rimraf.sync(dist)
     out.info('清理文档', dist)
-  } catch (error) {
-    // @ts-ignore
+  } catch (error: any) {
     out.err('清理文档', error.message)
   }
 }
@@ -27,8 +30,7 @@ export const cleanCache = (cachePath: string) => {
     const dist = path.join(__cwd, cachePath)
     fs.unlinkSync(dist)
     out.info('清理缓存', dist)
-  } catch (error) {
-    // @ts-ignore
+  } catch (error: any) {
     out.err('清理缓存', error.message)
   }
 }
@@ -41,8 +43,7 @@ export const cleanImages = (imgsPath: string) => {
     const dist = path.join(__cwd, imgsPath)
     rimraf.sync(dist)
     out.info('清理图片', dist)
-  } catch (error) {
-    // @ts-ignore
+  } catch (error: any) {
     out.err('清理图片', error.message)
   }
 }
